# backend/cmd/api/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Backend API built with Golang using Clean Code architecture. \\nGithub: [https://github.com/shion0625/backend].",
    "title": "E-commerce Application Backend API",
    "contact": {
      "name": "For API Support",
      "email": "nikhilnarayanan623@gmail.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "basePath": "/api",
  "paths": {
...
    "/admin/offers": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "API for admin to get all offers",
        "tags": ["Admin Offers"],
        "summary": "Get all offers (Admin)",
        "operationId": "GetAllOffers",
        "parameters": [
          {
            "type": "integer",
            "description": "Page Number",
            "name": "page_number",
            "in": "query"
          },
          {
            "type": "integer",
            "description": "Count",
            "name": "count",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully\tfound\tall\toffers",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          },
          "500": {
            "description": "Failed to get all offers",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          }
        }
      },
      "post": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "API for admin to add an offer (Admin)",
        "tags": ["Admin Offers"],
        "summary": "Add offer (Admin)",
        "operationId": "SaveOffer",
        "parameters": [
          {
            "description": "input field",
            "name": "input",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/request.Offer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully offer added",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          },
          "400": {
            "description": "Invalid inputs",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          },
          "409": {
            "description": "Offer already exist",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          }
        }
      }
    },
    "/admin/offers/category": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "API for admin to get all category offers",
        "tags": ["Admin Offers"],
        "summary": "Get all category offers (Admin)",
        "operationId": "GetAllCategoryOffers",
        "parameters": [
          {
            "type": "integer",
            "description": "Page Number",
            "name": "page_number",
            "in": "query"
          },
          {
            "type": "integer",
            "description": "Count",
            "name": "count",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "successfully got all offer_category",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          },
          "500": {
            "description": "failed to get offers_category",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          }
        }
      },
      "post": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "API for admin to add an offer category",
        "tags": ["Admin Offers"],
        "summary": "Add category offer (Admin)",
        "operationId": "SaveCategoryOffer",
        "parameters": [
          {
            "description": "input field",
            "name": "input",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/request.OfferCategory"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successfully offer added for category",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          },
          "400": {
            "description": "invalid input",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          }
        }
      },
      "patch": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "API admin to change product offer to another offer",
        "tags": ["Admin Offers"],
        "summary": "Change product offer (Admin)",
        "operationId": "ChangeCategoryOffer",
        "parameters": [
          {
            "description": "input field",
            "name": "input",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/request.UpdateCategoryOffer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successfully offer replaced for category",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          },
          "400": {
            "description": "invalid input",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          }
        }
      }
    },
    "/admin/offers/category/{offer_category_id}": {
      "delete": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "API admin to remove a offer from category",
        "tags": ["Admin Offers"],
        "summary": "Remove category offer (Admin)",
        "operationId": "RemoveCategoryOffer",
        "parameters": [
          {
            "type": "integer",
            "description": "Offer Category ID",
            "name": "offer_category_id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "successfully offer added for category",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          },
          "400": {
            "description": "invalid input",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          }
        }
      }
    },
    "/admin/offers/products": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "API for admin to get all product offers",
        "tags": ["Admin Offers"],
        "summary": "Get all product offers (Admin)",
        "operationId": "GetAllProductsOffers",
        "parameters": [
          {
            "type": "integer",
            "description": "Page Number",
            "name": "page_number",
            "in": "query"
          },
          {
            "type": "integer",
            "description": "Count",
            "name": "count",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "successfully got all offers_categories",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          },
          "500": {
            "description": "failed to get offer_products",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          }
        }
      },
      "post": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "API for admin to add an offer for product",
        "tags": ["Admin Offers"],
        "summary": "Add product offer (Admin)",
        "operationId": "SaveProductOffer",
        "parameters": [
          {
            "description": "input field",
            "name": "input",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/request.OfferProduct"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successfully offer added for product",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          },
          "400": {
            "description": "invalid input",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          }
        }
      },
      "patch": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "API admin to change product offer to another offer",
        "tags": ["Admin Offers"],
        "summary": "Change product offer (Admin)",
        "operationId": "ChangeProductOffer",
        "parameters": [
          {
            "description": "input field",
            "name": "input",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/request.UpdateProductOffer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully offer changed for  given product offer",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          },
          "400": {
            "description": "invalid input",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          }
        }
      }
    },
    "/admin/offers/products/{offer_product_id}": {
      "delete": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "API admin to remove a offer from product",
        "tags": ["Admin Offers"],
        "summary": "Remove product offer (Admin)",
        "operationId": "RemoveProductOffer",
        "parameters": [
          {
            "type": "integer",
            "description": "offer_product_id",
            "name": "offer_product_id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully offer removed from product",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          },
          "400": {
            "description": "invalid input on params",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          }
        }
      }
    },
    "/admin/offers/{offer_id}": {
      "delete": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "API admin to remove an offer",
        "tags": ["Admin Offers"],
        "summary": "Remove offer (Admin)",
        "operationId": "RemoveOffer",
        "parameters": [
          {
            "type": "integer",
            "description": "Offer ID",
            "name": "offer_id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "successfully offer added",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          },
          "400": {
            "description": "invalid input",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "request.Offer": {
      "type": "object",
      "required": [
        "description",
        "discount_rate",
        "end_date",
        "offer_name",
        "start_date"
      ],
      "properties": {
        "description": {
          "type": "string",
          "maxLength": 50,
          "minLength": 6
        },
        "discount_rate": {
          "type": "integer",
          "maximum": 100,
          "minimum": 1
        },
        "end_date": {
          "type": "string"
        },
        "offer_name": {
          "type": "string"
        },
        "start_date": {
          "type": "string"
        }
      }
    },
    "request.OfferCategory": {
      "type": "object",
      "required": ["category_id", "offer_id"],
      "properties": {
        "category_id": {
          "type": "integer"
        },
        "offer_id": {
          "type": "integer"
        }
      }
    },
    "request.OfferProduct": {
      "type": "object",
      "required": ["offer_id", "product_id"],
      "properties": {
        "offer_id": {
          "type": "integer"
        },
        "product_id": {
          "type": "integer"
        }
      }
    },
...
    "request.UpdateCategoryOffer": {
      "type": "object",
      "required": ["category_offer_id", "offer_id"],
      "properties": {
        "category_offer_id": {
          "type": "integer"
        },
        "offer_id": {
          "type": "integer"
        }
      }
    },
...
    "request.UpdateProductOffer": {
      "type": "object",
      "required": ["offer_id", "product_offer_id"],
      "properties": {
        "offer_id": {
          "type": "integer"
        },
        "product_offer_id": {
          "type": "integer"
        }
      }
    },
...
    "response.Response": {
      "type": "object",
      "properties": {
        "data": {},
        "error": {},
        "message": {
          "type": "string"
        },
        "success": {
          "type": "boolean"
        }
      }
    },
...
  },
  "securityDefinitions": {
    "BearerAuth": {
      "description": "Add prefix of Bearer before  token Ex: \"Bearer token\"",
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}